Add build task
Add a `build` task to perform local builds.

Include node modules
* "fs-extra" to recursively create, remove and copy entire directories.
* "ignore" to process the .tabrisignore file (gitignore style).

Use only synchronous node commands. As all steps have to be performed
sequentially, the extra complexity for async operations would not bring
any benefit.

To test the build command, run the CLI in a temporary test folder and
include a bin directory with `npm` and `cordova` mocks in the $PATH.

Change-Id: Ia4e45eb4bb3ce4d78ec3e5e2fcd1b7379ab3b8e0