Support request logging
Given the -l/--logging switch, successful and failed requests will be
logged to the console, e.g.:

* Success:
  GET /package.json

* Failure:
  GET /non-existant 404: "Not found"

With an error handler of the union server being configured, errors won't
be sent in the response message body. This prevents a problem on the
Tabris.js Windows client, which caused npm modules to take very long to
load.

Change-Id: Ibf24cd20f3d0362417d7ab1afecef6e822808d80