Fix failing tests under macOS
'temp' module creates temporary data under /var/folders. On macOS, this
is a symbolic link to /private/var/folders. The path provided by 'temp'
was the symbolic link at /var/folders/... pointing to
/private/var/folders/.... The actual command line arguments contained
the resolved absolute path, which resulted in a difference between
expected and actual path.

To solve this problem, always use absolute paths to temporary data in
tests.

Change-Id: Ia704b58b8d133f8df13ef1bb143919c5e6f593f2