Don't clean build/cordova on build
It is useful to update an app's JavaScript code or Cordova configuration
without rebuilding the platform and used plugins.

Since the introduction of "tabris clean", it is not needed to implicitly
clean build/cordova in "tabris build".

"tabris build" now skips execution of "cordova platform add" if a
platform has already been added, because "cordova platform add" will
fail for an existing platform.

Change "mkdirsSync" import to "ensureDirSync" [1]. "mkdirsSync" is an
alias to "ensureDirSync". It behaves like "mkdir -p" and the original
name makes the intention more clear.

[1]: https://github.com/jprichardson/node-fs-extra/blob/master/docs/ensureDir-sync.md

Change-Id: I0f222c01ca55d080f3513278496976443cfd9104