Copy app to www/app
Cordova places some files and directories (e.g. cordova.js, plugins/,
etc.) in the www/ directory. Those files can clash with the app.
Moreover, it turned out that remote apps load from the URL tab in the
developer console cannot access bundled cordova plugins when the app
lives in the same directory.

Separate the app from cordova-specific files by moving it to a
subfolder www/app/.

Rewrite the config.xml to include a <content> element that points to
the correct entrypoint. If the config.xml already contains a content
element, prefix it.

Change-Id: I64f883bb53d36c5dd533f7f29a9e8584125b194b