Update README
Add the clean command and new command line switches to the README.

Remove section on project goals. Most of them are implemented,
and they are not interesting for users of the project who read
the project description on GitHub or npmjs.com.

Also remove section on development, these things are explained
in the CONTRIBUTING.md. Clarify the part on `npm link` there.

Change-Id: Ie4284a403f1da42bac340f067fa3b58b1ac34918