Support platform download during build
Introduce TABRIS_BUILD_KEY environment variable, which contains an API
key for the Tabris.js build service. It can be retrieved from the
account settings page on tabrisjs.com [1].

Keep TABRIS_<platform>_PLATFORM support for backward compatibility.

Introduce installedTabrisVersion variable in TabrisProject which is
initialized when the Cordova project is created. It is used to download
a matching platform.

Use yauzl [2] for decompression. Yauzl is a standards-compliant unzip
library based on node's zlib.

Fix #21

[1]: https://staging.tabrisjs.com/settings/account
[2]: https://www.npmjs.com/package/yauzl

Change-Id: Ia2efb7f049fe5a4309b1a3ca157b39c40c879083