Simplify temporary directory and file creation
It is not necessary to wrap the temporary folder creation in a separate
module since temp's API is simple to use. It was confusing to have such
a module at all. Some tests did use this module and some didn't. Use
temp's API consistently instead.

Use the synchronous temp API to make tests easier to read.

Change-Id: Ie8fac5257c66883c66ea62518f877238ee677795