Simplify platform provision
Encapsulate platform provision in its own module. Prefer a "provider"
over a model class, because it better fits the scenario of making
something available than working with data.

The downloader module was renamed in provider in 0129a7e for simplicity
and because provision and download tasks are cohesive.

Add basic behavioral tests for the provider module.

Change-Id: I8f2a4b5a3138c17157b5e85f7c62827e8b558632