Show download progress bar
Android platform is currently ~2.4 MB, iOS platform ~7.8 MB big. To
indicate that a task is ongoing and the CLI has not hang up during the
donwnload, indicate progress using a progress bar. The progress bar
disappears on successful download to prevent polluting the log.

Rework downloadFile() to be a class extending Node's EventEmitter, since
Promises are currently not suited for progress tracking. Prefer using a
class over a function to be able to register event listeners before
calling downloadFile().

Change-Id: Iddd0ec3b6eccc8c91efd3eea389646ef6bc2c336