Include build key hash in built app
Make it possible to identify the build key that was used to build a
particular app.

Write a file build/cordova/www/build-key.sha256 containing a fingerprint
of the Tabris.js build key. This file is currently only written after a
platform has been downloaded by the CLI. It is not written when using
the TABRIS_<platform>_PLATFORM environment variable or if a matching
platform version already exists at ~/.tabris-cli/platforms.

Change-Id: I259055b0e57ff77875174af8f2216a276cf90042