Lessen tabris version check
The tabris version check prevents building tabris 2.0.0 apps
with a 2.0.1 tabris-cli. This version was meant to ensure the
correct cordova version, which is a dependency of tabris-cli.

However, since cordova versions won't change in patch versions,
this check appears too strict.

Lessen the tabris version check to validate only the major.minor
version of the installed tabris.

Bump to 2.0.2

Change-Id: I93ce53598232229fbd32586990ca3edad17d0fd8