Disable platform minor version check
The version check in the build command enforces that the native
platform has the same majori and minor version as the tabris CLI.

Having multiple tabris-cli version around for building new minor
versions seems like overkill. We have to think this through again.

To be able to build both, tabris 2.0.x and 2.1.0-dev versions,
with the latest tabris-cli, disable the minor version check.

Change-Id: I3a814d3a43967a31ecff56aa78a95e854d9d04c4