Implement watch option for tabris serve
When the -w/--watch option is given, the "watch" script given in the
package.json of the app will be executed instead of the "build" script
before the app is served.

In comparison to the "build" script, which is waited for to exit before
the app is served, "watch" is spawned asynchronously. This makes it
suitable for long running tasks, e.g. compilers, which rebuild the code
on file system change.

Change-Id: Iabf63a3795ee2f8b8367c5ab6d0e8b16d124e68b