Enable $IS_DEBUG for default builds
Running "tabris run" or "tabris build" without either the --debug or 
--release option resulted in $IS_DEBUG and $IS_RELEASE variables 
both being replaced with false in config.xml.

Set the debug option to true when neither --release nor --debug are 
given, since this is the default Cordova behavior for build
commands [1]. This enables $IS_DEBUG when both options are omitted.

[1]: https://github.com/apache/cordova-lib/blob/552584a0d5f6167d0816db1c865e8a3728eed624/spec/plugman/projects/android/cordova/lib/build.js#L51

Fix #29

Change-Id: I6b602a28b82311296542257412c9b8c94b9a6435