Improve output of integration tests
Previously the command status was often the first thing to be checked.
This resulted in test output which was not helpful in case of a failing
test, e.g. "Expected 1 to be 0".

Move result.status expectation to the bottom, since it produces least
helpful output.

Verify stderr output consistently each time the command result is
checked. Make it first expectation, since it produces most helpful
output.

Change-Id: If6856263ff9ac6dcc3133a30b2862b27e19902b5