Don't exclude build/ and cordova/ from subfolders
Previously directories named build/ and cordova/ were unexpectedly also
excluded from the app subdirectories before building the app.

Only exclude './build/' and './cordova/', i.e. the directories located
in the app root.

Previously it was checked whether the ignore rules applied to the
absolute file paths. Only apply them to the paths relative to the app
root.

This resolves an issue causing the embedded Cordova examples not to run
in the developer app.

Change-Id: If148b545028966e015130e6a811b1d567e707e80