Fix folders with leading slash not being ignored
node-ignore expects paths in the output format of path.join() or
path.relative() (see node-ignore#30). Thus, paths like './foobar' are
invalid. node-ignore handles empty string correctly as the root
directory, thus remove extra './' wrapping.

Change default ignores for directories in the root not to start with
leading dot, since this does not seem to be a valid syntax for
.gitignore anyway.

Change-Id: I15cce6aacb3bc80fa3fc4963c5e1b51be0ff95db