Fix app path on Windows
path.join() uses a backslash to join the URL segments on Windows. Widget
content source should be a path consisting of segments joined by a
forward slash. Introduce an own join method in ConfigXml which behaves
similarly to path.join() but uses forward slash as separator on all
platforms.

No test for this use case added, since Node.js' "path" module defines
its behavior dependent on the platform when the module is loaded and
thus cannot be mocked.

Fix #6

Change-Id: I2ac68b3488aea960c6e093060bfff5e5b9187f95