Support a Windows-only build option --arch
Passing "--archs" to Cordova is recommended for Windows builds.
Introduce a Windows-only property "--arch" delegating to Cordova's
"--archs" to make using this API easier.

Enforce usage of "--arch" to prevent unexpected build results.

Improve display of wrapped lines in the help output to ensure they
are aligned with the remaining lines.

Change-Id: I8bc6505b0c0364098deb0f910206564fcfb5c2b8