Implement tabris run --target and --list-targets
When "--taget <id>" is given, the app will be deployed on the target
device with this ID. This makes it possible to chose the device the app
should be deployed to when more than one devices/emulators are
running/connected.

Executing tabris run with the "--list-targets" option will list the
target IDs of all connected devices and emulators.

Add missing 'option' headers in README.

Change-Id: I510c14c8799cf9713d88605d4454b22e97f7e650