Workaround Node issue causing incorrect relative paths on Windows
Relative paths are not properly resolved on Windows when using
posix.relative() [1]. Use the os-specific relative() instead of
posix.relative() and replace Windows with POSIX separators as a
workaround. This should be safe, since '/' is not allowed in paths on
Windows.

[1]: https://github.com/nodejs/node/issues/13683#issuecomment-309014169

Change-Id: Id815ab9db1f179356e9a7700f624310657b9d6a1
(cherry picked from commit 322a8a266621c0952f081d51669d33f4391c0380)