Print tabris app console logs to CLI terminal
When sideloading a tabris app via HTTP, use a WebSocket connection to
print calls to the log methods of the client-side `console` object
on the server side terminal. This is accomplished by the following
steps:

* The CLI "serve" command starts a WebSocket server
* A "debug client" is concatenated to boot.js when it is requested
* That client is started by the Tabris.js module after initialization
* This establishes a "debug connection" to the WebSocket server
* The CLI closes any previous debug connections if they exit
* The debug client listens to "log" events and forwards them to the CLI

The messages sent via WebSocket are in a JSON format that is documented
in debug-protocol.md.

Unit tests are included to the degree that they are possible using a
mock WebSocket.

Change-Id: Id6f57d9ec430cd721ba000fb5b829c44c4e38f8c