Fix false test negatives on windows
Some tests are broken when running on windows. This change fixes the
majority of those where it is clearly the test implementation itself is
broken.

The remaining failing tests are likely either true negatives, or
not directly related to differences of the windows platform (unstable on
Linux and MacOS as well).

Change-Id: I544a76c563bd9f5333e53a2b4d6ecbc67039b385