Split path parameter in --project and --main options
The default "path" parameter used to do two very different things:

When pointed to a directory with a package.json, it would open the file
server in that directory, and run the build/watch scripts in the
package.json.

When pointed to a .js file, it would open the file server in the current
working directory and create a "synthetic" package.json with a "main"
field pointing to the given file. No build/watch scripts were run.

To allow serving a .js file while also running build/watch scripts,
split the default parameter in to the --project option to give the
project directory and the --main option to override the "main" field in
package.json.

Favor testing the server via integration tests in serve.test.js over
module tests in Server.test.js.

Change-Id: I4387883180c7e44521c2a29f684a1c6e2795cf2d