Revise failed platform download handling
This makes the unit test work on windows where file deletion does not
work as expected due to a file system quirk.

Instead of creating the platform file before starting the download and
deleting it on error, only create it when the download starts. In this
scenario the file deletion issue is entirely avoided.

In case the download aborts, download.js handles the file deletion. On
windows the tests use existsSync to check the file, which already
returns false. However, the file won't actually be deleted until node
exits. The prevents the temporary test directory from being deleted.
This issue can not be fixed and a warning is printed instead.

Restore temp file tracking in build.test.js, previously removed by
accident.

Change-Id: I95d30932cf14e95ba9d110a69a2def9e55a6fef5