Sent all buffered messages at once
Previously, buffered messages were removed after all messages had been
sent without a state check, which would sometimes cause messages to be
lost. With this change, all buffered messages are sent as an array if
the device parameter message is sent successfully. This buffer array
would be handled in `DebugConnection`.

The buffered messages sent as a JSON array format that is documented in
`debug-protocol.md`.

This also fixes the error "device parameter is null" that appeared when
a device parameter message wasn't sent yet by adding default parameter
values in `DebugConnection`.

Change-Id: I1beff8e978980b28a05b82eee1cf586fe3e0e7b7