Prevent tsc watcher to print the report to stdout directly
tsc watcher uses `process.stdout.write` to report the file change but
console UI only overrides the console print methods to prompt the input
so the output is messed up.

Since override of process.stdout.write is not possible, thus, `spawn`
stdio option 'pipe' is given to `proc.exec` method to prevent direct
use of process.stdout when `tsc watch` option is given. The output of
this child process is redirected to console.

Change-Id: Ib200b0cf30323402736ad00a1bcb0ad45cf9c13a