Wait for a response when the command is submitted
Previously, it was possible to send another command before the result
of the previous command is printed. With this change input prompt is
paused until a response is received or the device is disconnected.

After execution is finished on client side, debugClient sends another
message to enable prompt, this message is documented in
`debug-protocol.md`.

Change-Id: Ie13bb94cc54edc54a1559bde9da8bb00782a8b96