Improve sideloading performance by pre-caching source files
The GetFilesMiddleware class can deliver multiple source files per
request. It is used by the class ModulePreLoader delivered as part of
the debug client.

The ModulePreLoader hooks in to the module system to request all source
files of a directory the first time it is accessed. Non-source files are
not delivered, but an entry is made to mark them as existing. From then
on every time the module system attempts to access a file from
that directory it is either already cached, or requested only if it is
known to exist.

Server.js now forwards "deliver" events from GetFileMiddleware for
Watcher.js to consume. The modified package.json content is provided as
a file system overlay via FilesService.js.

Unlike other classes GetFilesMiddleware is tested not as a full
integration test (i.e. spawning a cli instance), but by hooking a
ModulePreloader instance up directly with its public API. This requires
some fixture to fake tabris and the http requests, but is overall much
faster and reliable. The Watcher tests still use it in a full
integration test.

Increased some timouts in unrelated tests to make the review build pass.

Change-Id: Ic001c63f140ce3ac8c6a6ebe940ed77437bd2eb3