Fix console commands not sent while waiting for pong
The websocket uses a ping-pong system to periodically check if the
connection is still alive. There is a short delay between the "ping" and
the "pong" where DebugConnection.js did set _isAlive to false, but this
did not necessarily mean the websocket was not usable. Therefore it was
not correct to reject input commands when _isAlive is false.

Replace _isAlive with the more expressive _waitingForPong. Only reject
commands when the websocket is actually closed. Should the connection
actually be dead the command may be lost, but DebugServer _onDisconnect
will mark the pending command as completed anyway.

Some refactoring including not keeping references to closed connections
in DebugServer.js and avoiding test-specific check in for websocket.ping
in DebugConnection.js.

Change-Id: I0b0f522ba7567fee98022dd1e9dd514993fed9f7