Prevent crash when readline fires keypress while prompt is disabled
This issue is difficult to replicate since it depends on what seems to
be a race condition in the readline module. However, if the user abuses
the keyoard too much while using the interactive remote console it is
possible that readline fires "keypress" events while the prompt is
disabled which causes an exception in Terminal#clearInput. This is
prevented by checking the promptEnabled property before Terminal
propagates the event.

Change-Id: I2fec0baf8218fc37d70fa1fb8daf21ed99c781b5