Prevent Cordova from being installed in unexpected location
When running `npm install`, npm tries to find a parent directory, which
contains either a `node_modules` folder or `package.json` and installs
the module relative to them [1]. This is a problem, since if a user has
`node_modules` folder in his home directory, the Cordova CLI will be
installed in `~/node_modules` instead of
`~/.tabris-cli/cordova/${version}/node_modules`.

To prevent that, create a valid package.json in the Cordova cache
directory, so that Cordova will always be installed in the expected
directory.

[1]: https://github.com/npm/npm/issues/2896#issuecomment-32741508

Change-Id: I54b219f242c64acbc8d293ee21547299a0007f31