Temporarily fix #67 by using commit ID dependency
The version 2.1.0 of ecstatic was unpublished from npm.
This caused an error when trying to npm install tabris-cli.
As result all builds > 1.11.0 at https://tabris.com started to fail
as they install the tabris-cli as part of the build.

We will publish another update with a "real" fix.
That is something equivalent of porting the tabris-cli to use
the latest version of ecstatic that is released on npm. But this
takes some time.

Change-Id: Ic8801fddbefbcff61124ff2179184f218e37f904