Add options to control QR code independently from actual address
The --external option allows a URL to be given as the only external
address, which will then also be encoded in the QR code.

The --port option allows to define the actual port of the HTTP server so
it maches the one given via --external.

With the --no-intro option the QR code is not printed to the console,
but still available via the HTML site.

This HTML site is tweaked slightly to include a link to the developer
app documentation.

Change-Id: I6b0e1cafa9d603679e14585158591d14432a3cee