Fix propagation of non-zero exit codes and stderr for watch scripts
If a watch script exits with a non-zero exit code, exit the CLI process
with 1.

Propagate watch stderr to the main process and sanitize it like stdout.

Test exec call parameters instead of passing the call through to exec
for simplicity. When spawning a child process, the temp module would
clean up the project after executing a test while the child process
is using the temp directory, resulting in an error.

Note: stdio configuration [null, 'pipe', null] is equivalent to 'pipe' [1].

Fix #76

[1]: https://nodejs.org/api/child_process.html#child_process_options_stdio

Change-Id: Ia8df174bb9a37fcc7c9e0cd26494fda7b876734d