Improve naming of spawn helper
Node.js has two distinctive APIs for starting child processes, spawn and
exec. Exec returns a buffer with the output of the child process and is
suitable for processes that return short outputs. Spawn returns streams
and is suitable for processes that produce large amounts of output. The
helper used in the CLI project uses the spawn API internally and behaves
largely like it, so name it "spawn" and not "exec" to prevent confusion.

Change-Id: Idcb58300cc9cdaadec9e4235a7cd7abcacd83063