Let scope of command evaluation be global
Previously, the scope of command evaluation was the
_handleServerMessage() method of RemoteConsole. Prevent tampering with
the remoteConsole instance by executing the command in global scope by
utilizing indirect eval evocation [1][2].

The old test wrongly assumed that retaining variable value among command
executions would mean that a command is executed in a local scope. The
test passed because the local scope of the _handleServerMessage() method
is only valid for a single command execution.

[1]: https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_Objects/eval#Description
[2]: http://perfectionkills.com/global-eval-what-are-the-options/#indirect_eval_call_theory

Change-Id: I211dbaae06803917c96aeaf4bb1d3f8e049d9335