Fix tabris serve crash when reloading during websocket disconnect
The check whether the debug websocket is still open so far only relied
on the ping-pong mechanism. However, this is not sufficient, as there
can be an almost 10 second window before this is properly detected.
(That assumes a very quick successful ping-pong after which the
connection immediately drops, then after 5 seconds another ping-pong
which is marked as failed after 5 seconds.)

Add explicit connection checks. The check is executed after a successful
pint-pong and before calling send. Also listen to the 'close' event of
websocket. (This by itself would not be sufficient as there is also a
"closing" state.)

Not tests since this would be very hard to test with so much fixture
that it would become meaningless.

Note that this is just a quickfix, as there is still the problem of
auto-reloads getting lost.

Fix #77

Change-Id: I2bbc2218dc178a2b0e4932a9ee770eacf356c452