Fix wrong session ID and unexpected disconnect log
A connection could close and open in a short timespan, leading to the
disconnect callback being called multiple times. This sometimes resulted
into an outdated timeout handler being unexpectedly executed after a
successful connection, leading to a "disconnected" message being printed
to the console, although a connection was already open.

Change-Id: I030a4e5dea157defb73fd490cff49e34977a8fa9