Introduce qrcode-renderer option
Make utf8 the default renderer, since the terminal renderer produces too
large QR codes that often require resizing of the Visual Studio Code
terminal to enable scanning.

The terminal renderer should be used when the UTF-8 QR code printed in
the terminal is not scannable due to unfavorable shape of the used font
characters.

Change-Id: Id610b7686a8e4c06f310e94a97bd96086c5d65a4