Explicitly differentiate unit, integration and acceptance tests
Previously, it was not immediately clear whether a test suite contained
integration or unit tests. For testing new functionality, it is often
not justified to write integration tests due to the complexity overhead
and overall brittleness compared to unit tests [1]. Make it clear which
tests are related to integration and acceptance, so that developers can
more easily make the choice whether they want to extend an existing
suite or create a new suite for their cases.

Make it clear whether a module is tested on its own (unit test*),
whether it is tested alongside with (some) of its dependencies
(integration test**) or whether a complete CLI command input/output is
tested (acceptance test).

* A unit test is also considered such when a module has dependencies on
  the environment (e.g. fs) that are not mocked during the test for
  pragmatic reasons.

** An integration test is considered such when a module uses other
   project modules instead of mocked versions of them.

[1]: https://martinfowler.com/bliki/TestPyramid.html

Change-Id: I44469afccfb5f8cdd8d6f4829520c4bf18669959