Improve watch mode usability
Previously, the server would be started in parallel to the long-running
watch script. This means the server would be available before the
compilation is finished. If a client connects before the compilation
finishes, it may get restarted multiple times due to changes in the dist
folder when the auto reload flag is used.

In watch mode, run the build script first and only then start the server
to ensure that the app source files are compiled before the server is
available. To prevent redundant compilation, watch scripts should skip
the initial build (e.g. for tsc, use the "incremental" build flag).

Change-Id: Ia11c387f82fd5169c7ec80edd5611cf2ed422f03