Reload app on reconnect
Ensure a reload command is executed when an app with the same session ID
reconnects.

Migrate DebugConnection callback functions to EventEmitter events to
reduce the overhead of maintaining two event mechanisms and to enable
multiple subscribers.

Remove AppReloader integration tests, since they were very brittle due
to dependence on internals. Replace them with unit tests that test the
AppReloader module alone and basic acceptance tests based on behavior
inspecting a real WebSocket connection.

Fix a bug in the reloader resulting in file watching not working when
the -p/--project flag is given to the CLI.

Remove unneeded passing of appPath parameters, it was available as a
property on the Server already. Make it public to make it available to
the AppReloader as well.

Change-Id: I7970c3ec62c8edfcbad567510fa5b25da10f962f