Reduce duplication of output formatting
The formatting/styling methods may not be a pure responsibility of
Terminal, but due to lack of DI and the resulting cumbersomeness of
dependency management and since Terminal is an information expert [1]
for this functionality, implement those methods there for pragmatic
reasons.

Log "command could not be sent" error as other messages. It should not
be logged to stderr since Tabris CLI does not fail.

[1] https://en.wikipedia.org/wiki/GRASP_(object-oriented_design)#Information_expert

Change-Id: I0fe4adb535eb458abccfb8074d0ffaadcee8e6b6