Improve rendering of info blocks
Show info blocks in boxes to make them stand out better from the rest of
the output.

Ignore ANSI codes when inspecting the output of standard streams. A
dependency of "boxen" uses ANSI codes to determine the terminal size on
some platforms.

waitFor... functions to be deduplicated in a follow-up change.

Change-Id: Ifce06a2fd7370cdef3d7866d5dbc37c40a4cd4e6