Fix preemptive WebSocket closure on iOS app restart
The WebSocket connection always got terminated shortly after restarting
the iOS app, leading to logs printed on app start never being delivered
to the remote console. The WebSocket got terminated just after the data
had been "successfully" sent, which resulted in the data never being
delivered.

The WebSocket implementation on iOS emitted a "close" event shortly
after terminate() was called, but this happened late enough so that a
new this._webSocket instance had been assigned for the new connection,
which resulted in the "close" listener being called and the just opened
WebSocket being closed again.

Instead of relying on the presence of this._webSocket to determine
whether to call the "close" listener or not, deregister the listener
when it's not needed.

Change-Id: I757c8a14c831b8b88070e84af384ce04eff703e8