Let preloader use relative paths on iOS when evaluating code
The Module.execute method requires a URL to associate the evaluated code
with for stack traces. In 3.4 this was changed to be the absolute path
of the file on the developer-machine to support the Android debugger.
However, on windows these paths are not valid URLs, which iOS can't
handle. Therefore revert to using the relative unix-like paths on iOS.

Change-Id: I8d8bd85570410d5c4b8475f384b4b5c0b19fd167