Fix unhandled RemoteConsole disposal error
Previously, an error would be thrown when the client prints an info
message about the connection being lost, because it attempts to send
this message when the remote console is disposed.

Indicate failure to send a message due to disposal by returning a false
result instead, since this does not end up being an unhandled error.

Change-Id: Ib59be09010a210a0746c626d0207c030bdc22a9d