Workaround clearTimeout breakage on iOS
throttle() assumes that a timeout callback will never be called after
its timeout gets cleared. Since this happens due to a bug in the
Tabris.js iOS client, the throttle callback will sometime get called
with a null context resulting in an error. Ensure the _flush() throttle
callback context by wrapping it in a lambda.

Workaround for eclipsesource/tabris-js/issues/2058