Fix source map support when bundling dependencies with own source maps
If a packager such as WebPack is used to bundle the app and its
dependencies in to one large file it may happen that there are multiple
"sourceMappingURL" comments in the same module. In this case the first
one was detected, which is usually the wrong one (pointing to the source
map  of a dependency). This change makes sure always the last one is
used.

This change requires the ES2020 feature "String.prototype.matchAll".
This is available in current tabris clients, but not node 10, which we
are currently use to run the CLI tests on Jenkins. Therefore the unit
tests use a polyfill for now.

Fix #82

Change-Id: I2351f324b54d1776fd67167808b8465c9533ebc4